feat(projects): add heartbeat icon to project social-icon row
Add a heart-pulse SVG to the social-icon set and render it next to the
GitHub icon on project pages when the project's git URL is a github.com
repo. Extract the existing getHeartbeatUrl helper into utils/heartbeat
so both the action-row button and the social-icon row can share it.
