+++ components/social-icons/index.tsx
@@ -7,8 +7,9 @@ import Twitter from './twitter.svg'
 import Nostr from './nostr.svg'
 import Zapstore from './zapstore.svg'
 import Web from './globe-solid.svg'
+import Heartbeat from './heartbeat.svg'
 
-// Icons taken from: https://simpleicons.org/ and https://fontawesome.com (globe)
+// Icons taken from: https://simpleicons.org/ and https://fontawesome.com (globe, heartbeat)
 
 const components = {
   mail: Mail,
@@ -20,6 +21,7 @@ const components = {
   nostr: Nostr,
   zapstore: Zapstore,
   website: Web,
+  heartbeat: Heartbeat,
 }
 
 const SocialIcon = ({ kind, href, size = 8 }) => {
